Jumping up to numpy 2.2.0, it looks like with old numpy has casted to int64 automatically, with np 2.2.0 we need to explicitly cast the _x_position. The reason why int64 is picked insted of uint, is that in case of reward function the sign is important.

+++ gym_super_mario_bros/smb_env.py
@@ -139,7 +139,7 @@ class SuperMarioBrosEnv(NESEnv):
     def _x_position(self):
         """Return the current horizontal position."""
         # add the current page 0x6d to the current x
-        return self.ram[0x6d] * 0x100 + self.ram[0x86]
+        return np.int64(self.ram[0x6d] * 0x100 + self.ram[0x86])
 
     @property
     def _left_x_position(self):
